spacenet-building-detection/code/requirements.txt
# List all dependencies for your SageMaker training job
# PyTorch and torchvision versions are usually pre-installed on SageMaker DL Containers.
# Specify other required libraries here.
rasterio
geopandas
numpy
pandas
torch
fiona
shapely
tqdm
scikit-learn
Pillow==9.5.0 # Specify an exact version to avoid issues
timm          # For the Vision Transformer model
torch_geometric
torch-scatter
torch-sparse

# If `torch_geometric` direct install fails due to CUDA/PyTorch version mismatches,
# you might need to find the specific wheel file from PyG's website (data.pyg.org/whl/)
# For example, for PyTorch 1.13.1 and CUDA 11.7, you might need:
# torch_geometric@https://data.pyg.org/whl/torch-1.13.1%2Bcu117/torch_geometric-2.2.0-cp39-cp39-linux_x86_64.whl
# Make sure the PyTorch version matches the framework_version in your PyTorch estimator (e.g., '1.13.1').